// src/models/community.ts
export class Community {
  _id: string;
  name: string;
  icon: string = "assets/icon/logoCircle.png";

  email: string;
  emergencyNumber:string;
  firstName: string;
  guid: string;
  imageKeyPath: string;
  imageUrl: string;
  lastName: string;
  members: any[];
  phone: string;
  resourcesJSON: string;
}